+++ src/client/components/localisation/darwin_robot/model/darwin_motor.ts
@@ -1,0 +1,13 @@
+import { observable } from 'mobx'
+
+export class DarwinMotor {
+  @observable public angle: number
+
+  public constructor(opts: DarwinMotor) {
+    Object.assign(this, opts)
+  }
+
+  public static of() {
+    return new DarwinMotor({ angle: 0 })
+  }
+}
