+++ src/client/components/localisation/model/vector.ts
@@ -1,0 +1,80 @@
+import { observable } from 'mobx'
+import { computed } from 'mobx'
+import { action } from 'mobx'
+
+export class Vector3 {
+  @observable public x: number
+  @observable public y: number
+  @observable public z: number
+
+  public constructor(x: number, y: number, z: number) {
+    this.x = x
+    this.y = y
+    this.z = z
+  }
+
+  public static of() {
+    return new Vector3(0, 0, 0)
+  }
+
+  @computed get length(): number {
+    return Math.sqrt(this.x * this.x + this.y * this.y + this.z * this.z)
+  }
+
+  @action
+  public set(x: number, y: number, z: number): Vector3 {
+    this.x = x
+    this.y = y
+    this.z = z
+    return this
+  }
+
+  @action
+  public clone(): Vector3 {
+    return new Vector3(this.x, this.y, this.z)
+  }
+
+  @action
+  public copy(v: Vector3): Vector3 {
+    this.x = v.x
+    this.y = v.y
+    this.z = v.z
+    return this
+  }
+
+  @action
+  public normalize(): Vector3 {
+    return this.divideScalar(this.length)
+  }
+
+  @action
+  public multiplyScalar(scalar: number): Vector3 {
+    this.x *= scalar
+    this.y *= scalar
+    this.z *= scalar
+    return this
+  }
+
+  @action
+  public divideScalar(scalar: number): Vector3 {
+    if (scalar !== 0) {
+      const invScalar = 1 / scalar
+      this.x *= invScalar
+      this.y *= invScalar
+      this.z *= invScalar
+    } else {
+      this.x = 0
+      this.y = 0
+      this.z = 0
+    }
+    return this
+  }
+
+  @action
+  public add(movement: Vector3): Vector3 {
+    this.x += movement.x
+    this.y += movement.y
+    this.z += movement.z
+    return this
+  }
+}
